requirements.txt
regex~=2023.10.3
cython==0.29.36
hdbscan==0.8.33
# dspy-ai==2.2.0
dspy-ai
arize-phoenix 
openinference-instrumentation-dspy 
opentelemetry-exporter-otlp